+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_options.hpp
@@ -76,7 +76,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   };
 
   /**
-   * @brief Optional parameters for ServiceClient::ListFilesSystems
+   * @brief Optional parameters for ServiceClient::ListFileSystemsSinglePageOptions
    */
   struct ListFileSystemsSinglePageOptions
   {
@@ -175,9 +175,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   };
 
   /**
-   * @brief Optional parameters for FileSystemClient::ListPaths
+   * @brief Optional parameters for FileSystemClient::ListPathsSinglePage
    */
-  struct ListDataLakePathsOptions
+  struct ListPathsSinglePageOptions
   {
     /**
      * @brief Context for cancelling long running operations.